Former-commit-id: ba0b00c6c14cca2a529d4d02e50888267d22af09

+++ mail/mkenv.sh
@@ -1,30 +1,53 @@
 #!/bin/sh
 
+. $HOME/myConfigs/gfw/get_apt_proxy.sh
+
 MAIL_CONFIG=$HOME/myConfigs/mail
-MUTT_CONFIG=$MAIL_CONFIG/mutt
-MUTT_HOME=$HOME/.mutt
 
-mkdir -p $MUTT_HOME
-mkdir -p $MUTT_HOME/cache
-mkdir -p $MUTT_HOME/sent
-mkdir -p $MUTT_HOME/temp
+echo "Install mutt? [Y/n]"
+read INPUT
 
-ln -sfnv $MUTT_CONFIG/muttrc.general $MUTT_HOME/muttrc.general
-ln -sfnv $MUTT_CONFIG/muttrc.imap.gmail $MUTT_HOME/muttrc.imap
-ln -sfnv $MUTT_CONFIG/muttrc.local $MUTT_HOME/muttrc.local
-ln -sfnv $MUTT_HOME/muttrc.imap $MUTT_HOME/muttrc
+if [ -z "$INPUT" ]; then
+  INPUT='Y'
+fi
 
-ln -sfnv $MUTT_CONFIG/signature $MUTT_HOME/sig
-ln -sfnv $MUTT_CONFIG/mailcap $MUTT_HOME/mailcap
-ln -sfnv $MUTT_CONFIG/mutt_get_imap_passwd.py $MUTT_HOME/mutt_get_imap_passwd.py
+if echo "$INPUT" | grep -iq "^y"; then
+  MUTT_CONFIG=$MAIL_CONFIG/mutt
+  MUTT_HOME=$HOME/.mutt
 
-ln -sfnv $MUTT_CONFIG/themes/jellybeans.muttrc $MUTT_HOME/theme.muttrc
+  MUTT_PACKAGE=$(dpkg -l|cut -d " " -f 3|grep mutt)
+  if [ -z "$MUTT_PACKAGE" ]; then
+    echo "Installing mutt..."
+    sudo apt-get $APT_PROXY install mutt mutt-patched w3m w3m-img motmuch notmuch-mutt
+    if [ "$?" -ne 0 ]; then
+      echo 'Install mutt failed, please check the output of apt-get.'
+      exit 1
+    fi
+    echo "Installing mutt...Done."
+  fi
 
-# goobook
-ln -sfnv $MAIL_CONFIG/goobook/goobookrc $HOME/.goobookrc
+  mkdir -p $MUTT_HOME
+  mkdir -p $MUTT_HOME/cache
+  mkdir -p $MUTT_HOME/sent
+  mkdir -p $MUTT_HOME/temp
 
-# msmtp
-ln -sfnv $MAIL_CONFIG/msmtp/msmtprc $HOME/.msmtprc
+  ln -sfnv $MUTT_CONFIG/muttrc.general $MUTT_HOME/muttrc.general
+  ln -sfnv $MUTT_CONFIG/muttrc.imap.gmail $MUTT_HOME/muttrc.imap
+  ln -sfnv $MUTT_CONFIG/muttrc.local $MUTT_HOME/muttrc.local
+  ln -sfnv $MUTT_HOME/muttrc.imap $MUTT_HOME/muttrc
+
+  ln -sfnv $MUTT_CONFIG/signature $MUTT_HOME/sig
+  ln -sfnv $MUTT_CONFIG/mailcap $MUTT_HOME/mailcap
+  ln -sfnv $MUTT_CONFIG/mutt_get_imap_passwd.py $MUTT_HOME/mutt_get_imap_passwd.py
+
+  ln -sfnv $MUTT_CONFIG/themes/jellybeans.muttrc $MUTT_HOME/theme.muttrc
+
+  # goobook
+  ln -sfnv $MAIL_CONFIG/goobook/goobookrc $HOME/.goobookrc
+
+  # msmtp
+  ln -sfnv $MAIL_CONFIG/msmtp/msmtprc $HOME/.msmtprc
+fi
 
 # sup
 SUP_CONFIG=$MAIL_CONFIG/sup
